requirements.txt
asgiref==3.6.0
Django==4.1.7
django-cors-headers==4.2.0
django-filter==22.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
Markdown==3.4.3
PyJWT==2.7.0
pytz==2023.2
sqlparse==0.4.3
